requirements.txt
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain_community>=0.0.1
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
openai>=1.3.0